cnn/requirements.txt
# python3.5
absl-py==0.1.13
astor==0.6.2
bleach==1.5.0
certifi==2018.1.18
coremltools==0.8
gast==0.2.0
grpcio==1.10.1
h5py==2.7.1
html5lib==0.9999999
Keras==2.0.6
Markdown==2.6.11
numpy==1.14.2
pandas==0.22.0
Pillow==5.1.0
protobuf==3.5.2.post1
python-dateutil==2.7.2
pytz==2018.3
PyYAML==3.12
scikit-learn==0.19.1
scipy==1.0.1
six==1.10.0
tensorboard==1.7.0
tensorflow==1.7.0
termcolor==1.1.0
Theano==1.0.1
Werkzeug==0.14.1
